requirements.txt
aiofiles==24.1.0
aiogram==3.14.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
attrs==24.2.0
certifi==2024.8.30
frozenlist==1.5.0
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
numpy==2.1.3
pillow==11.0.0
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
typing_extensions==4.12.2
yarl==1.17.1
